+++ source/blender/editors/lineart/lineart_intern.h
@@ -53,9 +53,9 @@ void *list_append_pointer_static_pool(struct LineartStaticMemPool *mph, ListBase
 void *list_pop_pointer_no_free(ListBase *h);
 void list_remove_pointer_item_no_free(ListBase *h, LinkData *lip);
 
-LineartStaticMemPoolNode *mem_new_static_pool(struct LineartStaticMemPool *smp, int size);
-void *mem_static_aquire(struct LineartStaticMemPool *smp, int size);
-void *mem_static_aquire_thread(struct LineartStaticMemPool *smp, int size);
+LineartStaticMemPoolNode *mem_new_static_pool(struct LineartStaticMemPool *smp, size_t size);
+void *mem_static_aquire(struct LineartStaticMemPool *smp, size_t size);
+void *mem_static_aquire_thread(struct LineartStaticMemPool *smp, size_t size);
 void *mem_static_destroy(LineartStaticMemPool *smp);
 
 void tmat_make_ortho_matrix_44d(double (*mProjection)[4],
